Move api client init to App context
