+++ apps/gitness/src/App.tsx
@@ -1,9 +1,7 @@
-import React from 'react'
 import { createBrowserRouter, Navigate, RouterProvider } from 'react-router-dom'
 import { QueryClientProvider } from '@tanstack/react-query'
 import { ThemeProvider, RootLayout } from '@harnessio/playground'
 import { TooltipProvider } from '@harnessio/canary'
-import { CodeServiceAPIClient } from '@harnessio/code-service-client'
 import { queryClient } from './framework/queryClient'
 import PipelineListPage from './pages/pipeline-list'
 import SignInPage from './pages/signin'
@@ -17,33 +15,7 @@ import PipelineEditPage from './pages/pipeline-edit/pipeline-edit'
 import { LandingPage } from './pages/landing-page'
 import { AppProvider } from './framework/context/AppContext'
 
-const BASE_URL_PREFIX = '/api/v1'
-
 export default function App() {
-  React.useEffect(() => {
-    new CodeServiceAPIClient({
-      urlInterceptor: (url: string) => `${BASE_URL_PREFIX}${url}`,
-      requestInterceptor: (request: Request): Request => {
-        // Retrieve the token from storage and add to headers if available
-        const token = localStorage.getItem('token')
-        if (token) {
-          const newRequest = request.clone()
-          newRequest.headers.set('Authorization', `Bearer ${token}`)
-          return newRequest
-        }
-        return request
-      },
-      responseInterceptor: (response: Response) => {
-        switch (response.status) {
-          case 401:
-            localStorage.removeItem('token')
-            window.location.href = '/signin'
-        }
-        return response
-      }
-    })
-  }, [])
-
   const router = createBrowserRouter([
     {
       path: '/',
